fix(token): Switch to local token validation by default
Before this release, the library used the token introspection endpoint
to validate the token by default. This has performance implications and
is not the recommended approach as Keycloak does not support opaque
tokens as of today

BREAKING CHANGE: The default value for  has changed to False. Refer to the documentation for details and instructions how to change it back